backend/pyproject.toml
[project]
name = "elyse-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix-otel>=0.9.2",
    "gcsfs>=2025.3.2",
    "litellm>=1.69.1",
    "llama-index>=0.12.35",
    "llama-index-core>=0.12.35",
    "llama-index-llms-litellm>=0.4.2",
    "llama-index-llms-openai>=0.3.38",
    "llama-index-utils-workflow>=0.3.2",
    "nest-asyncio>=1.6.0",
    "openinference-instrumentation-llama-index>=4.2.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
]
